Extract shared checklist parser and marble layout algorithm
- checklistUtils.ts: Single parseChecklistItems() replaces identical
  copies in CheckInModal and GoalForm. Handles all legacy formats.
- marbleLayout.ts: computeLayout(), getMarblePositions(), jarInnerWidth()
  extracted from MarbleJar component. Pure functions, now testable
  independently without rendering React.
- MarbleJar.tsx reduced to rendering only (imports from marbleLayout)
- GoalForm.tsx imports ChecklistItem type and parser from shared lib
