+++ src/app/components/MarbleJar.tsx
@@ -2,92 +2,7 @@
 
 import { useRef, useEffect, useMemo } from "react";
 import { playMarbleDropSound } from "@/lib/sounds";
-
-/* ── Marble colors for jar ── */
-const MARBLE_COLORS = [
-  "#FF6B6B", "#4ECDC4", "#FFE66D", "#A78BFA", "#FB923C",
-  "#34D399", "#F472B6", "#5B9BD5", "#FF9A9E", "#C084FC",
-];
-
-/* ── Jar inner width at a given y position (from SVG path shape) ── */
-/* Jar walls: bottom ~x=40-120, middle ~x=42-118, neck ~x=48-112.
-   Subtract 2*radius margin so marbles stay fully inside the walls. */
-function jarInnerWidth(y: number): number {
-  if (y >= 150) return 68; // very bottom, walls at ~42-118
-  if (y >= 120) return 70; // lower body
-  if (y >= 90) return 66;  // middle
-  if (y >= 60) return 56;  // upper body, narrowing
-  return 46;               // near neck
-}
-
-/**
- * Dynamically compute marble size and row layout so that exactly
- * `targetCount` marbles fill the jar. Starts with the largest marble
- * size that fits and shrinks until all marbles fit within the jar shape.
- */
-function computeLayout(targetCount: number) {
-  const CENTER = 80;
-  const BOTTOM = 160;
-  const MIN_Y = 42;
-
-  let r = 10;
-  let stepX: number;
-  let stepY: number;
-  let rowCaps: number[] = [];
-
-  // Shrink marble radius until targetCount fits inside the jar
-  while (r >= 4) {
-    stepX = Math.ceil(r * 2.15);
-    stepY = Math.ceil(r * 2.15);
-    rowCaps = [];
-    let total = 0;
-    let y = BOTTOM;
-    while (y >= MIN_Y && total < targetCount) {
-      const w = jarInnerWidth(y);
-      const cap = Math.max(1, Math.floor((w - r) / stepX) + 1);
-      rowCaps.push(cap);
-      total += cap;
-      y -= stepY;
-    }
-    if (total >= targetCount) break;
-    r -= 0.5;
-  }
-
-  // Trim excess from the last row so total === targetCount
-  let total = rowCaps.reduce((a, b) => a + b, 0);
-  while (total > targetCount && rowCaps.length > 0) {
-    const last = rowCaps[rowCaps.length - 1];
-    const excess = total - targetCount;
-    if (excess >= last) {
-      rowCaps.pop();
-      total -= last;
-    } else {
-      rowCaps[rowCaps.length - 1] -= excess;
-      total = targetCount;
-    }
-  }
-
-  return { radius: r, stepX: Math.ceil(r * 2.15), stepY: Math.ceil(r * 2.15), rowCaps, center: CENTER, bottom: BOTTOM };
-}
-
-function getMarblePositions(count: number, targetCount: number) {
-  const { radius, stepX, stepY, rowCaps, center, bottom } = computeLayout(targetCount);
-  const positions: { cx: number; cy: number }[] = [];
-  let placed = 0;
-
-  for (let row = 0; row < rowCaps.length && placed < count; row++) {
-    const cap = rowCaps[row];
-    const toPlace = Math.min(cap, count - placed);
-    const y = bottom - row * stepY;
-    for (let col = 0; col < toPlace; col++) {
-      const x = center + (col - (toPlace - 1) / 2) * stepX;
-      positions.push({ cx: x, cy: y });
-    }
-    placed += toPlace;
-  }
-
-  return { positions, radius };
-}
+import { getMarblePositions, MARBLE_COLORS } from "@/lib/marbleLayout";
 
 interface MarbleJarProps {
   successCount: number;
@@ -101,7 +16,6 @@ export default function MarbleJar({ successCount, targetCount }: MarbleJarProps)
     [successCount, targetCount]
   );
 
-  // Detect new marbles inside the effect to avoid stale closure issues
   const isNewMarble = prevCountRef.current !== null && successCount > prevCountRef.current;
 
   useEffect(() => {
